requirements.txt
Flask
requests
numpy
joblib
beautifulsoup4
gunicorn
pandas
xgboost
scikit-learn
SQLAlchemy
mysql-connector-python
gspread
google-auth
supabase
psycopg2-binary
python-dotenv